+++ src/event.rs
@@ -54,7 +54,9 @@ pub use self::{
         BlocklistDns, BlocklistDnsFields, CryptocurrencyMiningPool, CryptocurrencyMiningPoolFields,
         DnsCovertChannel, DnsEventFields, LockyRansomware,
     },
-    ftp::{BlocklistFtp, FtpBruteForce, FtpBruteForceFields, FtpEventFields, FtpPlainText},
+    ftp::{
+        BlocklistFtp, FtpBruteForce, FtpBruteForceFields, FtpCommand, FtpEventFields, FtpPlainText,
+    },
     http::{
         BlocklistHttp, BlocklistHttpFields, DgaFields, DomainGenerationAlgorithm, HttpEventFields,
         HttpThreat, HttpThreatFields, NonBrowser, RepeatedHttpSessions, RepeatedHttpSessionsFields,
@@ -2769,6 +2771,7 @@ mod tests {
 
     use chrono::{TimeZone, Utc};
 
+    use crate::FtpCommand;
     use crate::{
         Store,
         event::{
@@ -3947,6 +3950,18 @@ mod tests {
 
     #[tokio::test]
     async fn syslog_for_ftpplaintext() {
+        let command = FtpCommand {
+            command: "ls".to_string(),
+            reply_code: "200".to_string(),
+            reply_msg: "OK".to_string(),
+            data_passive: false,
+            data_orig_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 3)),
+            data_resp_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 4)),
+            data_resp_port: 10001,
+            file: "/etc/passwd".to_string(),
+            file_size: 5000,
+            file_id: "123".to_string(),
+        };
         let fields = FtpEventFields {
             src_addr: IpAddr::V4(Ipv4Addr::LOCALHOST),
             src_port: 10000,
@@ -3956,18 +3971,8 @@ mod tests {
             end_time: 100,
             user: "user1".to_string(),
             password: "password".to_string(),
-            command: "ls".to_string(),
-            reply_code: "200".to_string(),
-            reply_msg: "OK".to_string(),
-            data_passive: false,
-            data_orig_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 3)),
             sensor: "collector1".to_string(),
-            data_resp_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 4)),
-            data_resp_port: 10001,
-            file: "/etc/passwd".to_string(),
-            file_size: 5000,
-            file_id: "123".to_string(),
-            confidence: 1.0,
+            commands: vec![command],
             category: EventCategory::LateralMovement,
         };
 
@@ -3982,7 +3987,7 @@ mod tests {
         let (_, _, syslog_message) = message.unwrap();
         assert_eq!(
             &syslog_message,
-            r#"time="1970-01-01T01:01:01+00:00" event_kind="FtpPlainText" category="LateralMovement" sensor="collector1" src_addr="127.0.0.1" src_port="10000" dst_addr="127.0.0.2" dst_port="21" proto="6" end_time="100" user="user1" password="password" command="ls" reply_code="200" reply_msg="OK" data_passive="false" data_orig_addr="127.0.0.3" data_resp_addr="127.0.0.4" data_resp_port="10001" file="/etc/passwd" file_size="5000" file_id="123" confidence="1""#
+            r#"time="1970-01-01T01:01:01+00:00" event_kind="FtpPlainText" category="LateralMovement" sensor="collector1" src_addr="127.0.0.1" src_port="10000" dst_addr="127.0.0.2" dst_port="21" proto="6" end_time="100" user="user1" password="password" commands=[{command="ls", reply_code="200", reply_msg="OK", data_passive="false", data_orig_addr="127.0.0.3", data_resp_addr="127.0.0.4", data_resp_port="10001", file="/etc/passwd", file_size="5000", file_id="123"}]"#
         );
 
         let ftp_plain_text = Event::FtpPlainText(FtpPlainText::new(
@@ -3992,11 +3997,23 @@ mod tests {
         .to_string();
         assert_eq!(
             &ftp_plain_text,
-            r#"time="1970-01-01T01:01:01+00:00" event_kind="FtpPlainText" category="LateralMovement" sensor="collector1" src_addr="127.0.0.1" src_port="10000" dst_addr="127.0.0.2" dst_port="21" proto="6" end_time="100" user="user1" password="password" command="ls" reply_code="200" reply_msg="OK" data_passive="false" data_orig_addr="127.0.0.3" data_resp_addr="127.0.0.4" data_resp_port="10001" file="/etc/passwd" file_size="5000" file_id="123" triage_scores="""#
+            r#"time="1970-01-01T01:01:01+00:00" event_kind="FtpPlainText" category="LateralMovement" sensor="collector1" src_addr="127.0.0.1" src_port="10000" dst_addr="127.0.0.2" dst_port="21" proto="6" end_time="100" user="user1" password="password" commands=[{command="ls", reply_code="200", reply_msg="OK", data_passive="false", data_orig_addr="127.0.0.3", data_resp_addr="127.0.0.4", data_resp_port="10001", file="/etc/passwd", file_size="5000", file_id="123"}] triage_scores="""#
         );
     }
 
     fn ftpeventfields() -> FtpEventFields {
+        let command = FtpCommand {
+            command: "ls".to_string(),
+            reply_code: "200".to_string(),
+            reply_msg: "OK".to_string(),
+            data_passive: false,
+            data_orig_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 3)),
+            data_resp_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 4)),
+            data_resp_port: 10001,
+            file: "/etc/passwd".to_string(),
+            file_size: 5000,
+            file_id: "123".to_string(),
+        };
         FtpEventFields {
             src_addr: IpAddr::V4(Ipv4Addr::LOCALHOST),
             src_port: 10000,
@@ -4006,18 +4023,8 @@ mod tests {
             end_time: 100,
             user: "user1".to_string(),
             password: "password".to_string(),
-            command: "ls".to_string(),
-            reply_code: "200".to_string(),
-            reply_msg: "OK".to_string(),
-            data_passive: false,
-            data_orig_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 3)),
             sensor: "collector1".to_string(),
-            data_resp_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 4)),
-            data_resp_port: 10001,
-            file: "/etc/passwd".to_string(),
-            file_size: 5000,
-            file_id: "123".to_string(),
-            confidence: 1.0,
+            commands: vec![command],
             category: EventCategory::InitialAccess,
         }
     }
@@ -4037,7 +4044,7 @@ mod tests {
         let (_, _, syslog_message) = message.unwrap();
         assert_eq!(
             &syslog_message,
-            r#"time="1970-01-01T01:01:01+00:00" event_kind="BlocklistFtp" category="InitialAccess" sensor="collector1" src_addr="127.0.0.1" src_port="10000" dst_addr="127.0.0.2" dst_port="21" proto="6" end_time="100" user="user1" password="password" command="ls" reply_code="200" reply_msg="OK" data_passive="false" data_orig_addr="127.0.0.3" data_resp_addr="127.0.0.4" data_resp_port="10001" file="/etc/passwd" file_size="5000" file_id="123" confidence="1""#
+            r#"time="1970-01-01T01:01:01+00:00" event_kind="BlocklistFtp" category="InitialAccess" sensor="collector1" src_addr="127.0.0.1" src_port="10000" dst_addr="127.0.0.2" dst_port="21" proto="6" end_time="100" user="user1" password="password" commands=[{command="ls", reply_code="200", reply_msg="OK", data_passive="false", data_orig_addr="127.0.0.3", data_resp_addr="127.0.0.4", data_resp_port="10001", file="/etc/passwd", file_size="5000", file_id="123"}]"#
         );
 
         let blocklist_ftp = Event::Blocklist(RecordType::Ftp(BlocklistFtp::new(
@@ -4048,7 +4055,7 @@ mod tests {
 
         assert_eq!(
             &blocklist_ftp,
-            r#"time="1970-01-01T01:01:01+00:00" event_kind="BlocklistFtp" category="InitialAccess" sensor="collector1" src_addr="127.0.0.1" src_port="10000" dst_addr="127.0.0.2" dst_port="21" proto="6" end_time="100" user="user1" password="password" command="ls" reply_code="200" reply_msg="OK" data_passive="false" data_orig_addr="127.0.0.3" data_resp_addr="127.0.0.4" data_resp_port="10001" file="/etc/passwd" file_size="5000" file_id="123" triage_scores="""#
+            r#"time="1970-01-01T01:01:01+00:00" event_kind="BlocklistFtp" category="InitialAccess" sensor="collector1" src_addr="127.0.0.1" src_port="10000" dst_addr="127.0.0.2" dst_port="21" proto="6" end_time="100" user="user1" password="password" commands=[{command="ls", reply_code="200", reply_msg="OK", data_passive="false", data_orig_addr="127.0.0.3", data_resp_addr="127.0.0.4", data_resp_port="10001", file="/etc/passwd", file_size="5000", file_id="123"}] triage_scores="""#
         );
     }