+++ src/lib.rs
@@ -42,7 +42,24 @@ pub use self::cluster::*;
 pub use self::collections::Indexable;
 pub(crate) use self::collections::{IndexedMap, IndexedMapUpdate, IterableMap, Map};
 pub use self::column_statistics::*;
-pub use self::event::{Event, EventDb, EventKind, EventMessage};
+pub use self::event::{
+    BlocklistBootp, BlocklistBootpFields, BlocklistConn, BlocklistConnFields, BlocklistDceRpc,
+    BlocklistDceRpcFields, BlocklistDhcp, BlocklistDhcpFields, BlocklistDns, BlocklistDnsFields,
+    BlocklistFtp, BlocklistHttp, BlocklistHttpFields, BlocklistKerberos, BlocklistKerberosFields,
+    BlocklistLdap, BlocklistMqtt, BlocklistMqttFields, BlocklistNfs, BlocklistNfsFields,
+    BlocklistNtlm, BlocklistNtlmFields, BlocklistRdp, BlocklistRdpFields, BlocklistSmb,
+    BlocklistSmbFields, BlocklistSmtp, BlocklistSmtpFields, BlocklistSsh, BlocklistSshFields,
+    BlocklistTls, BlocklistTlsFields, CryptocurrencyMiningPool, CryptocurrencyMiningPoolFields,
+    DgaFields, Direction, DnsCovertChannel, DnsEventFields, DomainGenerationAlgorithm, Event,
+    EventDb, EventFilter, EventIterator, EventKind, EventMessage, ExternalDdos, ExternalDdosFields,
+    ExtraThreat, FilterEndpoint, FlowKind, FtpBruteForce, FtpBruteForceFields, FtpCommand,
+    FtpEventFields, FtpPlainText, HttpEventFields, HttpThreat, HttpThreatFields, LdapBruteForce,
+    LdapBruteForceFields, LdapEventFields, LdapPlainText, LearningMethod, LockyRansomware,
+    MultiHostPortScan, MultiHostPortScanFields, NetworkThreat, NetworkType, NonBrowser, PortScan,
+    PortScanFields, RdpBruteForce, RdpBruteForceFields, RecordType, RepeatedHttpSessions,
+    RepeatedHttpSessionsFields, SuspiciousTlsTraffic, TorConnection, TrafficDirection, TriageScore,
+    WindowsThreat, find_ip_country,
+};
 pub use self::migration::{migrate_backend, migrate_data_dir};
 pub use self::model::{Digest as ModelDigest, Model};
 pub use self::scores::Scores;